working with defineOutlet
