+++ templates/demo-store/src/App.server.tsx
@@ -10,20 +10,33 @@ import {
   ShopifyAnalytics,
   ShopifyProvider,
   CartProvider,
+  defineRSCOutlet,
 } from '@shopify/hydrogen';
 
 import {HeaderFallback} from '~/components';
 import type {CountryCode} from '@shopify/hydrogen/storefront-api-types';
 import {DefaultSeo, NotFound} from '~/components/index.server';
+import {Header} from './components/global/Header.server';
+import {Footer} from './components/global/Footer.server';
 
-function App({request, subRoute}: HydrogenRouteProps) {
+export const HeaderRSCOutlet = defineRSCOutlet({
+  outletName: 'HeaderRSCOutlet',
+  component: Header,
+});
+
+export const FooterRSCOutlet = defineRSCOutlet({
+  outletName: 'FooterRSCOutlet',
+  component: Footer,
+});
+
+function App({request, outlet}: HydrogenRouteProps) {
   const pathname = new URL(request.normalizedUrl).pathname;
   const localeMatch = /^\/([a-z]{2})(\/|$)/i.exec(pathname);
   const countryCode = localeMatch ? (localeMatch[1] as CountryCode) : undefined;
 
   const isHome = pathname === `/${countryCode ? countryCode + '/' : ''}`;
 
-  return subRoute ? (
+  return outlet ? (
     <ShopifyProvider countryCode={countryCode}>
       <FileRoutes basePath={countryCode ? `/${countryCode}/` : undefined} />
     </ShopifyProvider>
@@ -35,10 +48,21 @@ function App({request, subRoute}: HydrogenRouteProps) {
             <DefaultSeo />
           </Suspense>
           <Router>
-            <FileRoutes
-              basePath={countryCode ? `/${countryCode}/` : undefined}
-            />
-            <Route path="*" page={<NotFound />} />
+            <div className="flex flex-col min-h-screen">
+              <div className="">
+                <a href="#mainContent" className="sr-only">
+                  Skip to content
+                </a>
+              </div>
+              <HeaderRSCOutlet />
+              <main role="main" id="mainContent" className="flex-grow">
+                <FileRoutes
+                  basePath={countryCode ? `/${countryCode}/` : undefined}
+                />
+                <Route path="*" page={<NotFound />} />
+              </main>
+              <FooterRSCOutlet />
+            </div>
           </Router>
         </CartProvider>
         <PerformanceMetrics />
@@ -49,4 +73,7 @@ function App({request, subRoute}: HydrogenRouteProps) {
   );
 }
 
-export default renderHydrogen(App);
+export default renderHydrogen(App, {
+  HeaderRSCOutlet,
+  FooterRSCOutlet,
+});
