+++ templates/demo-store/src/components/global/index.server.ts
@@ -1,3 +1,2 @@
 export {Footer} from './Footer.server';
-export {Layout} from './Layout.server';
 export {NotFound} from './NotFound.server';
