+++ templates/demo-store/src/routes/account/reset/[id]/[resetToken].server.tsx
@@ -2,7 +2,6 @@ import {Suspense} from 'react';
 import {useRouteParams, Seo} from '@shopify/hydrogen';
 
 import {AccountPasswordResetForm} from '~/components';
-import {Layout} from '~/components/index.server';
 
 /**
  * This page shows a form for the user to enter a new password.
@@ -13,11 +12,11 @@ export default function ResetPassword() {
   const {id, resetToken} = useRouteParams();
 
   return (
-    <Layout>
+    <>
       <Suspense>
         <Seo type="noindex" data={{title: 'Reset password'}} />
       </Suspense>
       <AccountPasswordResetForm id={id} resetToken={resetToken} />
-    </Layout>
+    </>
   );
 }
