+++ templates/demo-store/src/routes/pages/[handle].server.tsx
@@ -10,7 +10,7 @@ import {
 import {Suspense} from 'react';
 
 import {PageHeader} from '~/components';
-import {NotFound, Layout} from '~/components/index.server';
+import {NotFound} from '~/components/index.server';
 
 export default function Page({params}: HydrogenRouteProps) {
   const {
@@ -37,7 +37,7 @@ export default function Page({params}: HydrogenRouteProps) {
   });
 
   return (
-    <Layout>
+    <>
       <Suspense>
         <Seo type="page" data={page} />
       </Suspense>
@@ -47,7 +47,7 @@ export default function Page({params}: HydrogenRouteProps) {
           className="prose dark:prose-invert"
         />
       </PageHeader>
-    </Layout>
+    </>
   );
 }
